Fix calendar start dates
